# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediERP - Medical Store ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory batches
      parameters:
        - name: productId
          in: query
          schema: { type: integer }
        - name: view
          in: query
          schema:
            type: string
            enum: [tablets, packs, boxes]
        - name: nearExpiry
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Inventory batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryBatch"
    post:
      operationId: receiveInventory
      tags: [inventory]
      summary: Receive new inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryReceiptInput"
      responses:
        "201":
          description: Inventory received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryBatch"
  /inventory/{id}:
    get:
      operationId: getInventoryBatch
      tags: [inventory]
      summary: Get an inventory batch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inventory batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryBatch"
    patch:
      operationId: updateInventoryBatch
      tags: [inventory]
      summary: Update inventory batch details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryBatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryBatch"
    delete:
      operationId: deleteInventoryBatch
      tags: [inventory]
      summary: Delete an inventory batch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryBatch"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: vendorId
          in: query
          schema: { type: integer }
        - name: accountId
          in: query
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense (creates journal entry)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── SALES / POS ───────────────────────────────────────────
...
components:
  schemas:
...
    InventoryBatch:
      type: object
      required: [id, productId, productName, unitType, boxesPurchased, packsPerBox, tabsPerPack, totalTablets, remainingTablets, costPerUnit, sellingPricePerUnit, expiryDate, receivedAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        batchNumber: { type: ["string", "null"] }
        unitType:
          type: string
        boxesPurchased: { type: integer }
        packsPerBox: { type: integer }
        tabsPerPack: { type: integer }
        totalTablets: { type: integer }
        remainingTablets: { type: integer }
        remainingPacks: { type: number }
        remainingBoxes: { type: number }
        costPerUnit: { type: number }
        sellingPricePerUnit: { type: number }
        sellingPricePerPack: { type: number }
        sellingPricePerBox: { type: number }
        expiryDate: { type: string, format: date }
        receivedAt: { type: string }
        vendorId: { type: ["integer", "null"] }
        vendorName: { type: ["string", "null"] }
        journalEntryId: { type: ["integer", "null"] }
    InventoryReceiptInput:
      type: object
      required: [productId, unitType, boxesPurchased, packsPerBox, tabsPerPack, costPerUnit, sellingPricePerUnit, expiryDate, paymentAccountId]
      properties:
        productId: { type: integer }
        batchNumber: { type: string }
        unitType:
          type: string
        boxesPurchased: { type: integer }
        packsPerBox: { type: integer }
        tabsPerPack: { type: integer }
        costPerUnit: { type: number }
        sellingPricePerUnit: { type: number }
        sellingPricePerPack: { type: number }
        sellingPricePerBox: { type: number }
        expiryDate: { type: string, format: date }
        vendorId: { type: integer }
        paymentAccountId: { type: integer }
        receivedAt: { type: string, format: date }
        notes: { type: string }
    InventoryBatchUpdate:
      type: object
      properties:
        batchNumber: { type: string }
        costPerUnit: { type: number }
        sellingPricePerUnit: { type: number }
        sellingPricePerPack: { type: number }
        sellingPricePerBox: { type: number }
        expiryDate: { type: string, format: date }
        vendorId: { type: integer }
...
    Expense:
      type: object
      required: [id, date, amount, description, expenseAccountId, paymentAccountId, createdAt]
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        amount: { type: number }
        description: { type: string }
        expenseAccountId: { type: integer }
        expenseAccountName: { type: string }
        paymentAccountId: { type: integer }
        paymentAccountName: { type: string }
        vendorId: { type: ["integer", "null"] }
        vendorName: { type: ["string", "null"] }
        journalEntryId: { type: ["integer", "null"] }
        reference: { type: ["string", "null"] }
        createdAt: { type: string }
    ExpenseInput:
      type: object
      required: [date, amount, description, expenseAccountId, paymentAccountId]
      properties:
        date: { type: string, format: date }
        amount: { type: number }
        description: { type: string }
        expenseAccountId: { type: integer }
        paymentAccountId: { type: integer }
        vendorId: { type: integer }
        reference: { type: string }

    # ── SALE ─────────────────────────────────────────────────
...