# lib/api-zod/src/generated/types/inventoryBatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediERP - Medical Store ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryBatch {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  batchNumber?: string | null;
  unitType: string;
  boxesPurchased: number;
  packsPerBox: number;
  tabsPerPack: number;
  totalTablets: number;
  remainingTablets: number;
  remainingPacks?: number;
  remainingBoxes?: number;
  costPerUnit: number;
  sellingPricePerUnit: number;
  sellingPricePerPack?: number;
  sellingPricePerBox?: number;
  expiryDate: Date;
  receivedAt: string;
  /** @nullable */
  vendorId?: number | null;
  /** @nullable */
  vendorName?: string | null;
  /** @nullable */
  journalEntryId?: number | null;
}